requirements.txt
Pillow==4.2.1
numpy==1.17.2
tqdm==4.36.1
tensorflow==1.12.0
pdf2image
pytesseract
sklearn
pyyaml
opencv-python
keras
simplejson